Fix an issue in opengl clipping and generalize to one riveqtpath rather than having one for qt5 and one for qt6
