+++ src/RiveQtQuickItem/renderer/riveqtopenglrenderer.cpp
@@ -17,380 +17,6 @@
 #include "renderer/riveqtopenglrenderer.h"
 #include "renderer/riveqtutils.h"
 
-RiveQtOpenGLPath::RiveQtOpenGLPath(const RiveQtOpenGLPath &rqp)
-{
-    m_path = rqp.m_path;
-    m_subPaths = rqp.m_subPaths;
-    m_pathSegmentsData = rqp.m_pathSegmentsData;
-    m_pathSegmentsOutlineData = rqp.m_pathSegmentsOutlineData;
-}
-
-RiveQtOpenGLPath::RiveQtOpenGLPath(rive::RawPath &rawPath, rive::FillRule fillRule)
-{
-    m_path.clear();
-    m_path.setFillRule(RiveQtUtils::riveFillRuleToQt(fillRule));
-
-    for (const auto &[verb, pts] : rawPath) {
-        switch (verb) {
-        case rive::PathVerb::move:
-            m_path.moveTo(pts->x, pts->y);
-            break;
-        case rive::PathVerb::line:
-            m_path.lineTo(pts->x, pts->y);
-            break;
-        case rive::PathVerb::cubic:
-            m_path.cubicTo(pts[0].x, pts[0].y, pts[1].x, pts[1].y, pts[2].x, pts[2].y);
-            break;
-        case rive::PathVerb::close:
-            m_path.lineTo(pts->x, pts->y);
-            m_path.closeSubpath();
-            m_isClosed = true;
-            break;
-        default:
-            break;
-        }
-    }
-
-    generateVertices();
-}
-
-void RiveQtOpenGLPath::rewind()
-{
-    m_pathSegmentsData.clear();
-    m_pathSegmentsOutlineData.clear();
-    m_subPaths.clear();
-    m_path.clear();
-}
-
-void RiveQtOpenGLPath::fillRule(rive::FillRule value)
-{
-    switch (value) {
-    case rive::FillRule::evenOdd:
-        m_path.setFillRule(Qt::FillRule::OddEvenFill);
-        break;
-    case rive::FillRule::nonZero:
-        m_path.setFillRule(Qt::FillRule::WindingFill);
-        break;
-    }
-}
-
-void RiveQtOpenGLPath::addRenderPath(RenderPath *path, const rive::Mat2D &m)
-{
-    if (!path) {
-        return;
-    }
-
-    RiveQtOpenGLPath *qtPath = static_cast<RiveQtOpenGLPath *>(path);
-    QMatrix4x4 matrix(m[0], m[2], 0, m[4], m[1], m[3], 0, m[5], 0, 0, 1, 0, 0, 0, 0, 1);
-
-    m_subPaths.emplace_back(SubPath(qtPath, matrix));
-
-    generateVertices();
-}
-
-QVector<QVector<QVector2D>> RiveQtOpenGLPath::toVertices() const
-{
-    return m_pathSegmentsData;
-}
-
-// TODO: Optimize, this gets called each frame for a stroke/path
-// there will be shortcuts and caching possible
-QVector<QVector<QVector2D>> RiveQtOpenGLPath::toVerticesLine(const QPen &pen) const
-{
-    qreal lineWidth = pen.widthF();
-    Qt::PenJoinStyle joinType = pen.joinStyle();
-    Qt::PenCapStyle capStyle = pen.capStyle();
-
-    QVector<QVector<QVector2D>> lineData;
-
-    if (m_pathSegmentsOutlineData.isEmpty()) {
-        return lineData;
-    }
-
-    for (QVector<QVector2D> pathData : m_pathSegmentsOutlineData) {
-        if (pathData.size() <= 1) {
-            continue;
-        }
-
-        QVector<QVector2D> lineDataSegment;
-
-        int count = pathData.count();
-        int endIndex = m_isClosed ? count : count - 1;
-
-        auto addMiterJoin = [&](const QVector2D &p1, const QVector2D &p2, const QVector2D &offset) {
-            lineDataSegment.append(p1 - offset);
-            lineDataSegment.append(p1 + offset);
-            lineDataSegment.append(p2 - offset);
-            lineDataSegment.append(p2 + offset);
-        };
-
-        auto addRoundJoin = [&](const QVector2D &p1, const QVector2D &p2, const QVector2D &offset, int numSegments) {
-            QVector2D prevPoint1 = p1 - offset;
-            QVector2D prevPoint2 = p1 + offset;
-
-            float angleStep = M_PI / numSegments;
-            float angle = atan2(offset.y(), offset.x());
-
-            for (int i = 0; i < numSegments; ++i) {
-                float newAngle = angle + angleStep * (i + 1);
-                QVector2D newOffset(cos(newAngle) * lineWidth / 2, sin(newAngle) * lineWidth / 2);
-                QVector2D newPoint1 = p1 - newOffset;
-                QVector2D newPoint2 = p1 + newOffset;
-
-                lineDataSegment.append(prevPoint1);
-                lineDataSegment.append(prevPoint2);
-                lineDataSegment.append(newPoint1);
-
-                lineDataSegment.append(newPoint1);
-                lineDataSegment.append(prevPoint2);
-                lineDataSegment.append(newPoint2);
-
-                prevPoint1 = newPoint1;
-                prevPoint2 = newPoint2;
-            }
-
-            lineDataSegment.append(prevPoint1);
-            lineDataSegment.append(prevPoint2);
-            lineDataSegment.append(p2 - offset);
-
-            lineDataSegment.append(p2 - offset);
-            lineDataSegment.append(prevPoint2);
-            lineDataSegment.append(p2 + offset);
-        };
-
-        auto addBevelJoin = [&](const QVector2D &p1, const QVector2D &p2, const QVector2D &offset) {
-            QVector2D p1_outer = p1 - offset;
-            QVector2D p1_inner = p1 + offset;
-            QVector2D p2_outer = p2 - offset;
-            QVector2D p2_inner = p2 + offset;
-
-            lineDataSegment.append(p1_outer);
-            lineDataSegment.append(p1_inner);
-            lineDataSegment.append(p2_inner);
-
-            lineDataSegment.append(p1_outer);
-            lineDataSegment.append(p2_inner);
-            lineDataSegment.append(p2_outer);
-        };
-
-        auto addCap = [&](const QVector2D &p, const QVector2D &offset, const QVector2D &normal, bool isStart) {
-            switch (capStyle) {
-            case Qt::PenCapStyle::FlatCap:
-                // No additional vertices needed for FlatCap
-                break;
-            case Qt::PenCapStyle::RoundCap: {
-                int numSegments = 20;
-                float angleStep = M_PI / numSegments;
-                float startAngle = atan2(offset.y(), offset.x());
-
-                QVector<QVector2D> capVertices;
-                QVector2D previousOffset(cos(startAngle) * lineWidth / 2, sin(startAngle) * lineWidth / 2);
-                for (int i = 0; i < numSegments; ++i) {
-                    float newAngle = startAngle + angleStep * (i + 1);
-                    QVector2D newOffset(cos(newAngle) * lineWidth / 2, sin(newAngle) * lineWidth / 2);
-
-                    capVertices.append(p);
-                    capVertices.append(p + previousOffset);
-                    capVertices.append(p + newOffset);
-
-                    previousOffset = newOffset;
-                }
-
-                if (isStart) {
-                    lineDataSegment = capVertices + lineDataSegment;
-                } else {
-                    lineDataSegment.append(capVertices);
-                }
-
-                break;
-            }
-            case Qt::PenCapStyle::SquareCap: {
-                QVector<QVector2D> capVertices;
-                capVertices.append(p - normal * (lineWidth / 2) + offset);
-                capVertices.append(p + normal * (lineWidth / 2) + offset);
-                capVertices.append(p - normal * (lineWidth / 2) - offset);
-                capVertices.append(p + normal * (lineWidth / 2) - offset);
-
-                if (isStart) {
-                    lineDataSegment = capVertices + lineDataSegment;
-                } else {
-                    lineDataSegment.append(capVertices);
-                }
-                break;
-            }
-            default:
-                // No additional vertices needed for other cap styles
-                break;
-            }
-        };
-
-        for (int i = 0; i < endIndex; ++i) {
-            const QVector2D p1 = pathData[i];
-            const QVector2D p2 = pathData[(i + 1) % count];
-            const QVector2D diff = p2 - p1;
-
-            QVector2D normal(-diff.y(), diff.x());
-            normal.normalize();
-
-            const QVector2D offset = normal * (lineWidth / 2.0);
-
-            if (!m_isClosed && (i == 0 || i == endIndex - 1)) {
-                addMiterJoin(p1, p2, offset);
-            } else {
-                switch (joinType) {
-                case Qt::PenJoinStyle::MiterJoin:
-                    addMiterJoin(p1, p2, offset);
-                    break;
-                case Qt::PenJoinStyle::RoundJoin:
-                    addRoundJoin(p1, p2, offset, 10);
-                    break;
-                case Qt::PenJoinStyle::BevelJoin:
-                    addBevelJoin(p1, p2, offset);
-                    break;
-                default:
-                    addMiterJoin(p1, p2, offset);
-                }
-            }
-        }
-
-        if (!m_isClosed) {
-            QVector2D startNormal(-pathData[1].y() + pathData[0].y(), pathData[1].x() - pathData[0].x());
-            startNormal.normalize();
-            QVector2D endNormal(-pathData.last().y() + pathData[count - 2].y(), pathData.last().x() - pathData[count - 2].x());
-            endNormal.normalize();
-
-            addCap(pathData.first(), startNormal * (lineWidth / 2.0), startNormal, true);
-            addCap(pathData.last(), -endNormal * (lineWidth / 2.0), endNormal, false);
-        }
-
-        lineData.append(lineDataSegment);
-    }
-
-    return lineData;
-}
-
-QPointF RiveQtOpenGLPath::cubicBezier(const QPointF &startPoint, const QPointF &controlPoint1, const QPointF &controlPoint2,
-                                      const QPointF &endPoint, qreal t)
-{
-    qreal oneMinusT = 1 - t;
-    qreal oneMinusTSquared = oneMinusT * oneMinusT;
-    qreal oneMinusTCubed = oneMinusTSquared * oneMinusT;
-    qreal tSquared = t * t;
-    qreal tCubed = tSquared * t;
-
-    QPointF point = oneMinusTCubed * startPoint + 3 * oneMinusTSquared * t * controlPoint1 + 3 * oneMinusT * tSquared * controlPoint2
-        + tCubed * endPoint;
-    return point;
-}
-
-QVector<QVector2D> RiveQtOpenGLPath::qpainterPathToVector2D(const QPainterPath &path)
-{
-    QVector<QVector2D> pathData;
-
-    if (path.isEmpty()) {
-        //        qCDebug(rqqpRendering) << "Path is empty, cannot calculate vertices";
-        return pathData;
-    }
-
-    QTriangleSet triangles = qTriangulate(path);
-
-    for (int i = 0; i < triangles.indices.size(); i++) {
-        int index;
-        if (triangles.indices.type() == QVertexIndexVector::UnsignedInt) {
-            index = static_cast<const quint32 *>(triangles.indices.data())[i];
-        } else {
-            index = static_cast<const quint16 *>(triangles.indices.data())[i];
-        }
-
-        qreal x = triangles.vertices[2 * index];
-        qreal y = triangles.vertices[2 * index + 1];
-
-        pathData.append(QVector2D(x, y));
-    }
-
-    return pathData;
-}
-
-QVector<QVector2D> RiveQtOpenGLPath::qpainterPathToOutlineVector2D(const QPainterPath &path)
-{
-    QVector<QVector2D> pathData;
-
-    if (path.isEmpty()) {
-        //        qCDebug(rqqpRendering) << "Path is empty, cannot calulate outline vector";
-        return pathData;
-    }
-
-    QPointF point = path.elementAt(0);
-    QVector2D centerPoint = QVector2D(point.x(), point.y());
-
-    // Add the center point of the triangle fan.
-    pathData.append(centerPoint);
-
-    for (int i = 1; i < path.elementCount(); ++i) {
-        QPainterPath::Element element = path.elementAt(i);
-
-        switch (element.type) {
-        case QPainterPath::MoveToElement:
-        case QPainterPath::LineToElement: {
-            QPointF point = element;
-            pathData.append(QVector2D(point.x(), point.y()));
-            break;
-        }
-
-        case QPainterPath::CurveToElement: {
-            QPointF startPoint = pathData.last().toPointF();
-            QPointF controlPoint1 = element;
-            QPointF controlPoint2 = path.elementAt(i + 1);
-            QPointF endPoint = path.elementAt(i + 2);
-
-            // Calculate the number of segments you want for the curve.
-            // You can change this value depending on the desired level of detail.
-            const int segments = 20;
-
-            for (int j = 1; j <= segments; ++j) {
-                qreal t = static_cast<qreal>(j) / segments;
-                QPointF point = cubicBezier(startPoint, controlPoint1, controlPoint2, endPoint, t);
-                pathData.append(QVector2D(point.x(), point.y()));
-            }
-
-            i += 2; // Skip the next two control points, as we already processed them.
-            break;
-        }
-        default:
-            break;
-        }
-    }
-
-    if (m_isClosed) {
-        // Connect the last point to the first point.
-        pathData.append(QVector2D(point.x(), point.y()));
-    }
-
-    return pathData;
-}
-
-void RiveQtOpenGLPath::generateVertices()
-{
-    if (m_path.isEmpty() && m_subPaths.empty()) {
-        qCDebug(rqqpRendering) << "Current path and sub paths are empty, cannot generate vertices.";
-        return;
-    }
-
-    m_pathSegmentsData.clear();
-    m_pathSegmentsOutlineData.clear();
-
-    m_pathSegmentsData.append(qpainterPathToVector2D(m_path));
-    m_pathSegmentsOutlineData.append(qpainterPathToOutlineVector2D(m_path));
-
-    for (SubPath &subPath : m_subPaths) {
-        QPainterPath sourcePath = subPath.path()->toQPainterPath();
-        QPainterPath transformedPath = RiveQtUtils::transformPathWithMatrix4x4(sourcePath, subPath.transform());
-
-        m_pathSegmentsData.append(qpainterPathToVector2D(transformedPath));
-        m_pathSegmentsOutlineData.append(qpainterPathToOutlineVector2D(transformedPath));
-    }
-}
-
 RiveQtOpenGLRenderer::RiveQtOpenGLRenderer()
     : rive::Renderer()
 {
@@ -716,7 +342,7 @@ void RiveQtOpenGLRenderer::drawPath(rive::RenderPath *path, rive::RenderPaint *p
         return;
     }
 
-    RiveQtOpenGLPath *qtPath = static_cast<RiveQtOpenGLPath *>(path);
+    RiveQtPath *qtPath = static_cast<RiveQtPath *>(path);
     RiveQtPaint *qtPaint = static_cast<RiveQtPaint *>(paint);
 
     const auto &blendMode = qtPaint->blendMode();
@@ -847,7 +473,7 @@ void RiveQtOpenGLRenderer::clipPath(rive::RenderPath *path)
     glClearColor(0.0f, 0.0f, 0.0f, 0.0f); // Clear color with alpha set to zero
     glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT | GL_STENCIL_BUFFER_BIT);
 
-    RiveQtOpenGLPath *qtPath = static_cast<RiveQtOpenGLPath *>(path);
+    RiveQtPath *qtPath = static_cast<RiveQtPath *>(path);
 
     QColor outlineColor(255, 0, 0, 255); // Red color for the outlines
 
@@ -1065,19 +691,4 @@ void RiveQtOpenGLRenderer::updateViewportSize()
 const QMatrix4x4 &RiveQtOpenGLRenderer::transform()
 {
     return m_Stack.back().transform;
-}
-
-SubPath::SubPath(RiveQtOpenGLPath *path, const QMatrix4x4 &transform)
-    : m_Path(path)
-    , m_Transform(transform)
-{
-}
-
-RiveQtOpenGLPath *SubPath::path() const
-{
-    return m_Path;
-}
-QMatrix4x4 SubPath::transform() const
-{
-    return m_Transform;
 }
